Add celbridge module with placeholder pypi info
