--- Celbridge/Workspace/Celbridge.Python/Assets/PythonExtras/cel/__init__.py
@@ -1,3 +1,0 @@
-from .cel_utils import run, install
-
-__all__ = ["run", "install"]
