+++ Celbridge/Workspace/Celbridge.Python/Assets/PythonExtras/celbridge/__init__.py
@@ -1,0 +1,2 @@
+from .customize_ipython import apply_ipython_customizations
+from .exit_lock import apply_exit_lock
