+++ Celbridge/Workspace/Celbridge.Python/Assets/PythonModules/celbridge/pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["hatchling>=1.25"]
+build-backend = "hatchling.build"
+
+[project]
+name = "celbridge"
+version = "0.1.0"
+description = "Example package showing a console script entry point for Celbridge."
+readme = "README.md"
+requires-python = ">=3.9"
+license = { text = "MIT" }
+authors = [{ name = "Your Name", email = "you@example.com" }]
+keywords = ["celbridge", "cli", "example"]
+classifiers = [
+  "Programming Language :: Python :: 3",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent",
+  "Topic :: Software Development :: Libraries",
+  "Topic :: Utilities"
+]
+
+# No third-party dependencies required; stdlib only
+dependencies = []
+
+[project.urls]
+Homepage = "https://example.com/celbridge"
+Repository = "https://github.com/example/celbridge"
+Issues = "https://github.com/example/celbridge/issues"
+
+[project.scripts]
+# This creates an executable named `celbridge` that calls celbridge.cli:main
+celbridge = "celbridge.cli:main"
+
+[project.optional-dependencies]
+ipython = ["ipython>=8"]
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/celbridge"]
+
+[tool.hatch.build.targets.sdist]
+include = ["src/celbridge", "README.md", "LICENSE", "pyproject.toml", "PUBLISHING.md"]
