+++ Celbridge/Workspace/Celbridge.Python/Assets/PythonModules/celbridge/src/celbridge/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: MIT
+
+from .customize_ipython import apply_ipython_customizations
+from .exit_lock import apply_exit_lock
+
+__version__ = "0.1.0"
+__all__ = ["apply_ipython_customizations", "apply_exit_lock"]
