+++ Celbridge/Workspace/Celbridge.Python/Assets/PythonModules/celbridge/src/celbridge/exit_lock.py
